+++ gym_donkeycar/__init__.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """Top-level package for OpenAI Gym Environments for Donkey Car."""
 import os
 
@@ -21,7 +19,7 @@ from gym_donkeycar.envs.donkey_env import (
 
 # Read version from file
 version_file = os.path.join(os.path.dirname(__file__), "version.txt")
-with open(version_file, "r") as file_handler:
+with open(version_file) as file_handler:
     __version__ = file_handler.read().strip()
 
 __author__ = """Tawn Kramer"""