+++ gym_donkeycar/core/sim_client.py
@@ -5,6 +5,9 @@ file: sim_client.py
 notes: wraps a tcp socket client with a handler to talk to the unity donkey simulator
 """
 import json
+from typing import Any, Dict, Tuple
+
+from gym_donkeycar.core.message import IMesgHandler
 
 from .client import SDClient
 
@@ -14,7 +17,7 @@ class SimClient(SDClient):
     Handles messages from a single TCP client.
     """
 
-    def __init__(self, address, msg_handler):
+    def __init__(self, address: Tuple[str, int], msg_handler: IMesgHandler):
         # we expect an IMesgHandler derived handler
         # assert issubclass(msg_handler, IMesgHandler)
 
@@ -27,29 +30,29 @@ class SimClient(SDClient):
         # we connect right away
         msg_handler.on_connect(self)
 
-    def send_now(self, msg):
+    def send_now(self, msg: Dict[str, Any]) -> None:
         # takes a dict input msg, converts to json string
         # and sends immediately. right now, no queue.
         json_msg = json.dumps(msg)
         super().send_now(json_msg)
 
-    def queue_message(self, msg):
+    def queue_message(self, msg: Dict[str, Any]) -> None:
         # takes a dict input msg, converts to json string
         # and adds to a lossy queue that sends only the last msg
         json_msg = json.dumps(msg)
         self.send(json_msg)
 
-    def on_msg_recv(self, jsonObj):
+    def on_msg_recv(self, json_obj: Dict[str, Any]) -> None:
         # pass message on to handler
-        self.msg_handler.on_recv_message(jsonObj)
+        self.msg_handler.on_recv_message(json_obj)
 
-    def is_connected(self):
+    def is_connected(self) -> bool:
         return not self.aborted
 
-    def __del__(self):
+    def __del__(self) -> None:
         self.close()
 
-    def close(self):
+    def close(self) -> None:
         # Called to close client connection
         self.stop()