+++ gym_donkeycar/core/util.py
@@ -3,15 +3,15 @@ import re
 import numpy as np
 
 
-def replace_float_notation(string):
+def replace_float_notation(string: str) -> str:
     """
     Replace unity float notation for languages like
     French or German that use comma instead of dot.
     This convert the json sent by Unity to a valid one.
     Ex: "test": 1,2, "key": 2 -> "test": 1.2, "key": 2
 
-    :param string: (str) The incorrect json string
-    :return: (str) Valid JSON string
+    :param string: The incorrect json string
+    :return: Valid JSON string
     """
     regex_french_notation = r'"[a-zA-Z_]+":(?P<num>[0-9,E-]+),'
     regex_end = r'"[a-zA-Z_]+":(?P<num>[0-9,E-]+)}'