+++ gym_donkeycar/envs/donkey_env.py
@@ -5,6 +5,7 @@ date: 2018-08-31
 """
 import logging
 import time
+from typing import Any, Callable, Dict, List, Optional, Tuple
 
 import gym
 import numpy as np
@@ -17,15 +18,25 @@ from gym_donkeycar.envs.donkey_sim import DonkeyUnitySimContoller
 logger = logging.getLogger(__name__)
 
 
-def supply_defaults(conf):
+def supply_defaults(conf: Dict[str, Any]) -> None:
+    """
+    Update the config dictonnary
+    with defaults when values are missing.
+
+    :param conf: The user defined config dict,
+        passed to the environment constructor.
+    """
     defaults = [
         ("start_delay", 5.0),
-        ("max_cte", 5.0),
+        ("max_cte", 8.0),
         ("frame_skip", 1),
         ("cam_resolution", (120, 160, 3)),
         ("log_level", logging.INFO),
         ("host", "localhost"),
         ("port", 9091),
+        ("steer_limit", 1.0),
+        ("throttle_min", 0.0),
+        ("throttle_max", 1.0),
     ]
 
     for key, val in defaults:
@@ -37,18 +48,17 @@ def supply_defaults(conf):
 class DonkeyEnv(gym.Env):
     """
     OpenAI Gym Environment for Donkey
+
+    :param level: name of the level to load
+    :param conf: configuration dictionary
     """
 
     metadata = {"render.modes": ["human", "rgb_array"]}
 
-    ACTION_NAMES = ["steer", "throttle"]
-    STEER_LIMIT_LEFT = -1.0
-    STEER_LIMIT_RIGHT = 1.0
-    THROTTLE_MIN = 0.0
-    THROTTLE_MAX = 1.0
-    VAL_PER_PIXEL = 255
+    ACTION_NAMES: List[str] = ["steer", "throttle"]
+    VAL_PER_PIXEL: int = 255
 
-    def __init__(self, level, conf=None):
+    def __init__(self, level: str, conf: Optional[Dict[str, Any]] = None):
         print("starting DonkeyGym env")
         self.viewer = None
         self.proc = None
@@ -62,7 +72,7 @@ class DonkeyEnv(gym.Env):
         supply_defaults(conf)
 
         # set logging level
-        logging.basicConfig(level=conf["log_level"])  # pytype: disable=key-error
+        logging.basicConfig(level=conf["log_level"])
 
         logger.debug("DEBUG ON")
         logger.debug(conf)
@@ -80,10 +90,12 @@ class DonkeyEnv(gym.Env):
         # start simulation com
         self.viewer = DonkeyUnitySimContoller(conf=conf)
 
+        # Note: for some RL algorithms, it would be better to normalize the action space to [-1, 1]
+        # and then rescale to proper limtis
         # steering and throttle
         self.action_space = spaces.Box(
-            low=np.array([self.STEER_LIMIT_LEFT, self.THROTTLE_MIN]),
-            high=np.array([self.STEER_LIMIT_RIGHT, self.THROTTLE_MAX]),
+            low=np.array([-float(conf["steer_limit"]), float(conf["throttle_min"])]),
+            high=np.array([float(conf["steer_limit"]), float(conf["throttle_max"])]),
             dtype=np.float32,
         )
 
@@ -94,49 +106,53 @@ class DonkeyEnv(gym.Env):
         self.seed()
 
         # Frame Skipping
-        self.frame_skip = conf["frame_skip"]  # pytype: disable=key-error
+        self.frame_skip = conf["frame_skip"]
 
-        # wait until loaded
+        # wait until the car is loaded in the scene
         self.viewer.wait_until_loaded()
 
-    def __del__(self):
+    def __del__(self) -> None:
         self.close()
 
-    def close(self):
+    def close(self) -> None:
         if hasattr(self, "viewer") and self.viewer is not None:
             self.viewer.quit()
         if hasattr(self, "proc") and self.proc is not None:
             self.proc.quit()
 
-    def set_reward_fn(self, reward_fn):
+    def set_reward_fn(self, reward_fn: Callable) -> None:
         self.viewer.set_reward_fn(reward_fn)
 
-    def set_episode_over_fn(self, ep_over_fn):
+    def set_episode_over_fn(self, ep_over_fn: Callable) -> None:
         self.viewer.set_episode_over_fn(ep_over_fn)
 
-    def seed(self, seed=None):
+    def seed(self, seed: Optional[int] = None):
         self.np_random, seed = seeding.np_random(seed)
         return [seed]
 
-    def step(self, action):
-        for i in range(self.frame_skip):
+    def step(self, action: np.ndarray) -> Tuple[np.ndarray, float, bool, Dict[str, Any]]:
+        for _ in range(self.frame_skip):
             self.viewer.take_action(action)
             observation, reward, done, info = self.viewer.observe()
         return observation, reward, done, info
 
-    def reset(self):
+    def reset(self) -> np.ndarray:
+        # Activate hand brake, so the car does not move
+        self.viewer.handler.send_control(0, 0, 1.0)
+        time.sleep(0.1)
         self.viewer.reset()
+        self.viewer.handler.send_control(0, 0, 1.0)
+        time.sleep(0.1)
         observation, reward, done, info = self.viewer.observe()
-        time.sleep(1)
         return observation
 
-    def render(self, mode="human", close=False):
+    def render(self, mode: str = "human", close: bool = False) -> Optional[np.ndarray]:
         if close:
             self.viewer.quit()
 
         return self.viewer.render(mode)
 
-    def is_game_over(self):
+    def is_game_over(self) -> bool:
         return self.viewer.is_game_over()
 
 
@@ -145,54 +161,54 @@ class DonkeyEnv(gym.Env):
 
 class GeneratedRoadsEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(GeneratedRoadsEnv, self).__init__(level="generated_road", *args, **kwargs)
+        super().__init__(level="generated_road", *args, **kwargs)
 
 
 class WarehouseEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(WarehouseEnv, self).__init__(level="warehouse", *args, **kwargs)
+        super().__init__(level="warehouse", *args, **kwargs)
 
 
 class AvcSparkfunEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(AvcSparkfunEnv, self).__init__(level="sparkfun_avc", *args, **kwargs)
+        super().__init__(level="sparkfun_avc", *args, **kwargs)
 
 
 class GeneratedTrackEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(GeneratedTrackEnv, self).__init__(level="generated_track", *args, **kwargs)
+        super().__init__(level="generated_track", *args, **kwargs)
 
 
 class MountainTrackEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(MountainTrackEnv, self).__init__(level="mountain_track", *args, **kwargs)
+        super().__init__(level="mountain_track", *args, **kwargs)
 
 
 class RoboRacingLeagueTrackEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(RoboRacingLeagueTrackEnv, self).__init__(level="roboracingleague_1", *args, **kwargs)
+        super().__init__(level="roboracingleague_1", *args, **kwargs)
 
 
 class WaveshareEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(WaveshareEnv, self).__init__(level="waveshare", *args, **kwargs)
+        super().__init__(level="waveshare", *args, **kwargs)
 
 
 class MiniMonacoEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(MiniMonacoEnv, self).__init__(level="mini_monaco", *args, **kwargs)
+        super().__init__(level="mini_monaco", *args, **kwargs)
 
 
 class WarrenTrackEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(WarrenTrackEnv, self).__init__(level="warren", *args, **kwargs)
+        super().__init__(level="warren", *args, **kwargs)
 
 
 class ThunderhillTrackEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(ThunderhillTrackEnv, self).__init__(level="thunderhill", *args, **kwargs)
+        super().__init__(level="thunderhill", *args, **kwargs)
 
 
 class CircuitLaunchEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(CircuitLaunchEnv, self).__init__(level="circuit_launch", *args, **kwargs)
+        super().__init__(level="circuit_launch", *args, **kwargs)