+++ HISTORY.rst
@@ -2,6 +2,17 @@
 History
 =======
 
+1.3.0 (WIP)
+------------------
+* Dropped Python 3.6 support, pinned Gym to version 0.21
+* Move steer limits and throttle limits to config dict
+* Normalized reward and use squared error for CTE
+* Enabled hand brake in ``send_control()`` and at reset time
+* Added type hints to most core methods
+* Added ``send_lidar_config()`` method to configure LIDAR
+* Added car roll, pitch yaw angle
+* Renamed lidar config to use snake case instead of CamelCase (for instance ``degPerSweepInc`` was renamed to ``deg_per_sweep_inc``)
+
 1.1.1 (2021-02-28)
 ------------------
 * Fix type checking error